+++ src/eval.ml
@@ -113,7 +113,9 @@ let rec eval
         (* drop it on the floor *)
         | Some snapshot ->
           let data = Snapshot.apply_action snapshot ~schedule_event action in
-          Map.set model ~key:id ~data
+          if model_info.equal data model_info.default
+          then Map.remove model id
+          else Map.set model ~key:id ~data
     in
     let%mapn apply_action = apply_action
     and result = results_map in
@@ -125,26 +127,21 @@ let rec eval
     let (T { t; model = model_info; action = action_info }) = Map.find_exn out_of key in
     let chosen_model =
       Incremental.map model ~f:(fun map ->
-        let (Enum_types.Case_model.T { model; info; t_of_sexp = _ }) =
-          Enum_types.Multi_model.find_exn map key
+        let (Hidden.Model.T { model; info; t_of_sexp = _ }) =
+          Hidden.Multi_model.find_exn map key
         in
         let equal = Type_equal.Id.same_witness_exn info.type_id model_info.type_id in
         Type_equal.conv equal model)
     in
-    let inject action =
-      inject (Enum_types.Case_action.T { action; type_id = action_info; key })
-    in
+    let inject action = inject (Hidden.Action.T { action; type_id = action_info; key }) in
     let%mapn snapshot = eval environment chosen_model ~inject t
     and model = model in
-    let apply_action
-          ~schedule_event
-          (Enum_types.Case_action.T { action; type_id; key = key' })
-      =
+    let apply_action ~schedule_event (Hidden.Action.T { action; type_id; key = key' }) =
       match key_equal key' key, Type_equal.Id.same_witness type_id action_info with
       | true, Some T ->
         let new_model = Snapshot.apply_action snapshot ~schedule_event action in
-        let new_model = Enum_types.Case_model.create model_info new_model in
-        Enum_types.Multi_model.set model ~key ~data:new_model
+        let new_model = Hidden.Model.create model_info new_model in
+        Hidden.Multi_model.set model ~key ~data:new_model
       | _ ->
         let key = sexp_of_key key in
         let action = Type_equal.Id.to_sexp type_id action in
@@ -155,6 +152,29 @@ let rec eval
               (key : Sexp.t)
               (action : Sexp.t)];
         model
+    in
+    Snapshot.create ~apply_action ~result:(Snapshot.result snapshot)
+  | Lazy lazy_computation ->
+    let (T { t; model = model_info; action = action_info }) = force lazy_computation in
+    let input_model =
+      let%map model = model in
+      let (Hidden.Model.T { model; info; _ }) =
+        Option.value model ~default:(Hidden.Model.create model_info model_info.default)
+      in
+      let witness = Type_equal.Id.same_witness_exn info.type_id model_info.type_id in
+      Type_equal.conv witness model
+    in
+    let inject action =
+      inject (Hidden.Action.T { action; type_id = action_info; key = () })
+    in
+    let%map snapshot = eval environment input_model ~inject t
+    and model = model in
+    let apply_action ~schedule_event (Hidden.Action.T { action; type_id; key = () }) =
+      match Type_equal.Id.same_witness type_id action_info with
+      | Some T ->
+        let new_model = Snapshot.apply_action snapshot ~schedule_event action in
+        Some (Hidden.Model.create model_info new_model)
+      | None -> model
     in
     Snapshot.create ~apply_action ~result:(Snapshot.result snapshot)
   | Wrap { model_id; inject_id; inner; apply_action } ->
@@ -189,4 +209,14 @@ let rec eval
           Snapshot.apply_action inner_snapshot ~schedule_event action2
         in
         outer_model, new_inner_model)
+  | With_model_resetter { t; default_model } ->
+    let reset_event = inject (First ()) in
+    let inject a = inject (Second a) in
+    let%map snapshot = eval environment model ~inject t in
+    Snapshot.create
+      ~result:(Snapshot.result snapshot, reset_event)
+      ~apply_action:(fun ~schedule_event action ->
+        match action with
+        | First () -> default_model
+        | Second a -> Snapshot.apply_action ~schedule_event snapshot a)
 ;;